PROYECTO2/backend_python/requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
redis